3단계 - 로또(2등) (#2975)
* refactor: `InputControl`의 하위 클래스에서 생성자 호출로 `View`를 주입하도록 변경

* feat(input): 보너스 볼 입력 구현

* refactor(input): 보너스 번호 입력을 `Number`로 바로 변환

* feat(counter): 보너스 번호를 포함하는 2등 로또를 세는 기능 추가

* fix(reward): 로또 상금의 총합이 `int` 범위를 넘어갈 수 있으므로 `long`으로 변경

* feat(output): 당첨 통계에 2등 관련 정보 추가

* fix(input): 입력 순서 바뀐 것 수정

* refactor: 리뷰 반영

+++ src/main/java/lotto/model/LottoRewardStatistics.java
@@ -12,15 +12,15 @@ public class LottoRewardStatistics {
 
 
     public float profit() {
-        int rewards = winnerRewards();
+        long rewards = winnerRewards();
         int spent = counter.total() * Lotto.PRICE;
 
         return rewards / (float) spent;
     }
 
-    public int winnerRewards() {
+    public long winnerRewards() {
         return Place.winners().stream()
-                .mapToInt(this::reward)
+                .mapToLong(this::reward)
                 .sum();
     }
